requirements.txt
playwright
python-dotenv
pydantic
rich
tenacity